+++ src/merenda/kosmo/application.nim
@@ -64,6 +64,7 @@ const
   KosmoMoeBottomAreaRows = 1
   KosmoTabIdentifierPrefix = "kosmo.buffer."
   KosmoHelpTabIdentifier* = "kosmo.help"
+  KosmoConfigTabIdentifier* = "kosmo.config"
   KosmoTerminalIdentifierPrefix = "kosmo.terminal."
   KosmoFilesTabIdentifier* = "kosmo.sidebar.files"
   KosmoFindTabIdentifier* = "kosmo.sidebar.find"